feat: separate course management pages
